+++ src/process.c
@@ -966,7 +966,7 @@ Proc         Status   Buffer         Command\n\
 	  tem = Fcar (Fcdr (p->status));
 	  if (XFASTINT (tem))
 	    {
-	      sprintf (tembuf, " %d", XFASTINT (tem));
+	      sprintf (tembuf, " %ld", (long) XFASTINT (tem));
 	      write_string (tembuf, -1);
 	    }
 	}
